move struct to primitives
