+++ bin/cliain/Cargo.toml
@@ -16,7 +16,6 @@ env_logger = "0.8"
 ink_metadata = { version = "3.0", features = ["derive"] }
 log = "0.4"
 pallet-staking = { git = "https://github.com/Cardinal-Cryptography/substrate.git", branch = "aleph-v0.9.23" }
-pallet-elections = { path = "../../pallets/elections" }
 primitives = { path = "../../primitives" }
 serde = { version = "1.0.137", features = ["derive"] }
 serde_json = "1.0.81"
