+++ e2e-tests/src/rewards.rs
@@ -6,9 +6,9 @@ use aleph_client::{
     RewardPoint, SessionKeys, SignedConnection,
 };
 use log::info;
-use pallet_elections::{CommitteeSeats, LENIENT_THRESHOLD};
+use pallet_elections::LENIENT_THRESHOLD;
 use pallet_staking::Exposure;
-use primitives::{EraIndex, SessionIndex};
+use primitives::{CommitteeSeats, EraIndex, SessionIndex};
 use sp_core::H256;
 use sp_runtime::Perquintill;
 use substrate_api_client::{AccountId, XtStatus};
