+++ e2e-tests/src/test/validators_rotate.rs
@@ -4,7 +4,7 @@ use aleph_client::{
     change_validators, get_current_session, wait_for_finalized_block, wait_for_full_era_completion,
     wait_for_session, AnyConnection, Header, KeyPair, RootConnection, SignedConnection,
 };
-use pallet_elections::CommitteeSeats;
+use primitives::CommitteeSeats;
 use sp_core::Pair;
 use substrate_api_client::{AccountId, XtStatus};
