+++ pallets/elections/Cargo.toml
@@ -7,7 +7,6 @@ license = "Apache 2.0"
 [dependencies]
 codec = { package = "parity-scale-codec", version = "3.0", default-features = false, features = ["derive"] }
 scale-info = { version = "2.0", default-features = false, features = ["derive"] }
-serde = { version = "1.0", optional = true }
 
 frame-support = { default-features = false, git = "https://github.com/Cardinal-Cryptography/substrate.git", branch = "aleph-v0.9.23" }
 frame-system = { default-features = false, git = "https://github.com/Cardinal-Cryptography/substrate.git", branch = "aleph-v0.9.23" }
@@ -40,6 +39,5 @@ std = [
     "primitives/std",
     "pallet-balances/std",
     "sp-staking/std",
-    "serde",
 ]
 try-runtime = ["frame-support/try-runtime"]
