+++ primitives/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "primitives"
-version = "0.4.0"
+version = "0.4.1"
 authors = ["Cardinal Cryptography"]
 edition = "2021"
 license = "Apache 2.0"
